Add RefreshToken grant
Create RefreshTokenGrant class, which authenticates the client, checks
that the refresh token is valid and checks that the refresh token was
issued to the same client that sent the request

Change the respond method in the BearerReponse class to accept a string
in the refreshToken variable. In that case, the refresh token sent to
the method will be forwarded into the reponse.